+++ lib/service/FirebaseMessaging.dart
@@ -119,12 +119,12 @@ class FirebaseMessaging extends rokwire.FirebaseMessaging with NotificationsList
   static const String notifySettingsFavoritesNotification              = "$notifyBase.settings.favorites";
   static const String notifySettingsAssessmentsNotification            = "$notifyBase.settings.assessments";
   static const String notifySettingsCalendarNotification               = "$notifyBase.settings.calendar";
-  static const String notifySettingsAppointmentsNotification        = "$notifyBase.settings.appointments";
-  static const String notifySettingsMapsNotification                    = "$notifyBase.settings.maps";
-  static const String notifySettingsContactsNotification               = "$notifyBase.settings.contacts";
-  static const String notifySettingsResearchNotification              = "$notifyBase.settings.research";
-  static const String notifySettingsPrivacyNotification                 = "$notifyBase.settings.privacy";
-  static const String notifySettingsNotificationsNotification         = "$notifyBase.settings.notifications";
+  static const String notifySettingsAppointmentsNotification           = "$notifyBase.settings.appointments";
+  static const String notifySettingsMapsNotification                   = "$notifyBase.settings.maps";
+  static const String notifySettingsAboutNotification                  = "$notifyBase.settings.contacts";
+  static const String notifySettingsResearchNotification               = "$notifyBase.settings.research";
+  static const String notifySettingsPrivacyNotification                = "$notifyBase.settings.privacy";
+  static const String notifySettingsNotificationsNotification          = "$notifyBase.settings.notifications";
   static const String notifyGuideArticleDetailNotification             = "$notifyBase.guide.article.detail";
 
   // Topic names
@@ -274,7 +274,7 @@ class FirebaseMessaging extends rokwire.FirebaseMessaging with NotificationsList
   static const String payloadTypeSettingsCalendar = 'settings.calendar';
   static const String payloadTypeSettingsAppointments = 'settings.appointments';
   static const String payloadTypeSettingsMaps = 'settings.maps';
-  static const String payloadTypeSettingsContacts = 'settings.contacts';
+  static const String payloadTypeSettingsAbout = 'settings.contacts';
   static const String payloadTypeSettingsResearch = 'settings.research';
   static const String payloadTypeSettingsNotifications = 'settings.notifications';
   static const String payloadTypeSettingsPrivacy = 'settings.privacy';
@@ -629,8 +629,8 @@ class FirebaseMessaging extends rokwire.FirebaseMessaging with NotificationsList
     else if (type == payloadTypeSettingsMaps) {
       NotificationService().notify(notifySettingsMapsNotification, data);
     }
-    else if (type == payloadTypeSettingsContacts) {
-      NotificationService().notify(notifySettingsContactsNotification, data);
+    else if (type == payloadTypeSettingsAbout) {
+      NotificationService().notify(notifySettingsAboutNotification, data);
     }
     else if (type == payloadTypeSettingsResearch) {
       NotificationService().notify(notifySettingsResearchNotification, data);